refactor(services): convert reply service to pure factory function
