+++ src/services/reply/index.ts
@@ -1,4 +1,2 @@
-import { ReplyService } from "./ReplyService";
-import { ReplyServiceQueries } from "./ReplyServiceQueries";
-
-export { ReplyService, ReplyServiceQueries };
+export { createReplyService } from './service';
+export { createReplyQuery } from './query';
